@@ -3,43 +3,43 @@
   "version": "0.0.1",
   "description": "Mosquito Alert anomaly detection",
   "productName": "Anomaly Detection Map",
-  "author": "Enric Pou <e.pou@ceab.csic.es>",
+  "author": "Guillermo Sanz <guillermo.sanz@ceab.csic.es>",
+  "type": "module",
   "private": true,
   "scripts": {
-    "lint": "eslint --ext .js,.vue ./",
-    "format": "prettier --write \"**/*.{js,vue,scss,html,md,json}\" --ignore-path .gitignore",
+    "lint": "eslint -c ./eslint.config.js \"./src*/**/*.{ts,js,cjs,mjs,vue}\"",
+    "format": "prettier --write \"**/*.{js,ts,vue,scss,html,md,json}\" --ignore-path .gitignore",
     "test": "echo \"No test specified\" && exit 0",
     "dev": "quasar dev",
     "build": "quasar build",
-    "deploy": "sh deploy.sh"
+    "postinstall": "quasar prepare"
   },
   "dependencies": {
     "@quasar/extras": "^1.16.4",
     "axios": "^1.2.1",
-    "echarts": "^5.5.1",
-    "ol": "^10.2.1",
-    "ol-contextmenu": "^5.5.0",
-    "ol-ext": "^4.0.24",
     "quasar": "^2.16.0",
     "vue": "^3.4.18",
-    "vue-echarts": "^7.0.3",
-    "vue-i18n": "^9.0.0",
-    "vue-router": "^4.0.12",
-    "vue3-openlayers": "^11.2.1"
+    "vue-i18n": "^11.0.0",
+    "vue-router": "^4.0.12"
   },
   "devDependencies": {
-    "@intlify/vite-plugin-vue-i18n": "^3.3.1",
-    "@quasar/app-vite": "^1.9.0",
+    "@eslint/js": "^9.14.0",
+    "@vue/eslint-config-typescript": "^14.4.0",
+    "@vue/eslint-config-prettier": "^10.1.0",
+    "@types/node": "^20.5.9",
+    "@intlify/unplugin-vue-i18n": "^4.0.0",
+    "@quasar/app-vite": "^2.1.0",
     "autoprefixer": "^10.4.2",
-    "eslint": "^8.57.0",
-    "eslint-config-prettier": "^8.1.0",
-    "eslint-plugin-vue": "^9.0.0",
-    "postcss": "^8.4.14",
-    "prettier": "^2.5.1",
-    "vite-plugin-checker": "^0.6.4"
+    "eslint": "^9.14.0",
+    "eslint-plugin-vue": "^9.30.0",
+    "globals": "^15.12.0",
+    "prettier": "^3.3.3",
+    "typescript": "~5.5.3",
+    "vue-tsc": "^2.0.29",
+    "vite-plugin-checker": "^0.9.0"
   },
   "engines": {
-    "node": "^20 || ^18 || ^16",
+    "node": "^28 || ^26 || ^24 || ^22 || ^20 || ^18",
     "npm": ">= 6.13.4",
     "yarn": ">= 1.21.1"
   }
