@@ -4,4 +4,4 @@
 export default {
   failed: 'Action failed',
   success: 'Action was successful'
-}
+};
